minor fix in policy-io.c, seaudit - recent files, error checking on open files
2003/11/06 13:17:18-00:00 !mayerf
Fixed a couple of bugs relating to open options in libapol.

2003/10/15 19:24:50-00:00 !mayerf
Implemented changes to support optional policy load capabilities.

2003/10/10 18:23:23-00:00 !mayerf
Coomprehensive update that primarily implements the creation of two libseuser libs; one core C lib and one that adds the TCL support.  This required extensive changes to the seuser command line program, and the entire removal of the wrapper layer.  It also had ancillary impact on libapol, and on the make process for all tools.

2003/10/08 16:04:47-00:00 !mayerf
Added stubs for optionsal loading of pieces of policies.

2003/10/03 14:27:20-00:00 !mayerf
Added missing policy-io.* files.

(Logical change 1.3)


git-svn-id: http://oss.tresys.com/repos/setools/trunk@166 c5f68610-3bfa-0310-bcc3-b3c1dee5e5c1