Bug fix in libapol parse code. Change to correctly return the index of the newly added rule.
Also, removed incrementing of policy rule count. This is already handled in policy.c. Incrementing here
gives an inaccurate policy rule count number. The rule count should be incremented within the policy
code and not the policy parse code. This may need to be checked for other rules types as well.

(Logical change 1.239)


git-svn-id: http://oss.tresys.com/repos/setools/trunk/setools@1127 c5f68610-3bfa-0310-bcc3-b3c1dee5e5c1