fix merge problems
2005/01/13 04:20:23-05:00 (none)!donp
cleanup

2005/01/13 03:11:23-05:00 (none)!donp
Changed to const strings.

2005/01/07 15:21:29-05:00 tresys.com!bwhalen
Change the api to use 2d char arrays

2005/01/03 14:30:16-05:00 tresys.com!bwhalen
Moved defines to allow for stat64 and to allow replcon to use sefs_get_file_class

2004/12/30 14:59:47-05:00 tresys.com!bwhalen
removed a bunch of defines to .c file
rewrote header to print_valid_object_classes

2004/12/30 10:23:15-05:00 tresys.com!bwhalen
Moved headers to .c file that are no longer part of api

2004/12/29 14:42:25-05:00 tresys.com!bwhalen
Put in header to close a db

2004/12/23 10:49:12-05:00 tresys.com!bwhalen
Added in some new dfines, changed the api of get_types to allow user to get
all known entities.
Added api to deal with char** arrays.

2004/12/21 14:56:29-05:00 tresys.com!bwhalen
Comments plus changed api of functions handling returns

2004/12/21 13:29:27-05:00 tresys.com!bwhalen
Put in a new struct for returned data from searching, also put in api's for
a print, and destroy on that data struct.

2004/12/20 16:19:52-05:00 tresys.com!bwhalen
Edited teh api to reflect our new hidden "db"

2004/12/17 14:35:11-05:00 tresys.com!bwhalen
Added in some new structs to handle passing of database and for searches,
and moved most of the api to .c file to hide it from functions just including
the library

(Logical change 1.514)


git-svn-id: http://oss.tresys.com/repos/setools/trunk/setools@2205 c5f68610-3bfa-0310-bcc3-b3c1dee5e5c1